feat: gateway service implemented

+++ user-management-service/src/main/java/com/bit/usermanagementservice/service/UserServiceImpl.java
@@ -1,8 +1,10 @@
 package com.bit.usermanagementservice.service;
 
 import com.bit.usermanagementservice.UserManagementServiceApplication;
+import com.bit.usermanagementservice.dto.AuthUserRequest;
 import com.bit.usermanagementservice.dto.UserRequest;
 import com.bit.usermanagementservice.dto.UserResponse;
+import com.bit.usermanagementservice.exception.InvalidRoleException;
 import com.bit.usermanagementservice.exception.UserNotFoundException;
 import com.bit.usermanagementservice.exception.UserNotSoftDeletedException;
 import com.bit.usermanagementservice.model.AppUser;
@@ -12,6 +14,8 @@ import com.bit.usermanagementservice.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 import java.util.ArrayList;
 import java.util.List;
@@ -21,9 +25,12 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class UserServiceImpl implements UserService {
 
+    @Value("#{'${default-roles}'.split(', ')}")
+    private final List<String> DEFAULT_ROLES;
     private static final Logger logger = LogManager.getLogger(UserManagementServiceApplication.class);
     private final UserRepository userRepository;
     private final RoleRepository roleRepository;
+    private final GatewayService gatewayService;
 
     @Override
     public UserResponse getUser(Long id) {
@@ -56,6 +63,11 @@ public class UserServiceImpl implements UserService {
     public UserResponse createUser(UserRequest userRequest) {
         logger.info("Creating a user");
 
+        validateRoles(userRequest.getRoles());
+        checkUniqueness(userRequest);
+
+        gatewayService.createUser(mapToAuthUserRequest(userRequest));
+
         AppUser user = AppUser.builder()
                 .name(userRequest.getName())
                 .username(userRequest.getUsername())
@@ -70,11 +82,17 @@ public class UserServiceImpl implements UserService {
         return mapToUserResponseWithoutPassword(user);
     }
 
+
     @Override
     public UserResponse updateUser(Long id, UserRequest userRequest) {
         logger.info("Updating user with ID {}: {}", id, userRequest);
+
+        validateRoles(userRequest.getRoles());
+
         AppUser existingUser = userRepository.findById(id)
                 .orElseThrow(() -> new UserNotFoundException("User doesn't exist with id " + id));
+
+        checkUniquenessForUpdate(existingUser, userRequest);
 
         existingUser.setName(userRequest.getName());
         existingUser.setUsername(userRequest.getUsername());
@@ -82,7 +100,10 @@ public class UserServiceImpl implements UserService {
         existingUser.setPassword(userRequest.getPassword());
         existingUser.setRoles(getRolesAsRole(userRequest.getRoles()));
 
+
         userRepository.save(existingUser);
+
+        gatewayService.updateUser(id, mapToAuthUserRequest(userRequest));
 
         logger.info("User updated with id {}", id);
         return mapToUserResponseWithoutPassword(existingUser);
@@ -96,24 +117,64 @@ public class UserServiceImpl implements UserService {
         userRepository.restoreUser(id);
         AppUser user = userRepository.findById(id)
                 .orElseThrow(() -> new UserNotFoundException("Couldn't restore the user with id " + id));
+
+        gatewayService.restoreUser(id);
+
         return mapToUserResponseWithoutPassword(user);
     }
 
     @Override
     public void deleteUser(Long id) {
         logger.info("Deleting user with ID: {}", id);
-        AppUser existingUser = userRepository.findById(id)
-                .orElseThrow(() -> new UserNotFoundException("User not found with id " + id));
+
+        if(!userRepository.existsById(id))
+           throw new UserNotFoundException("User not found with id " + id);
 
         userRepository.deleteById(id);
-        logger.info("Deleted user: {}", existingUser);
+
+        gatewayService.deleteUser(id);
+
+        logger.info("User soft-deleted");
     }
 
     @Override
     public void deleteUserPermanently(Long id) {
+        if(!userRepository.existsById(id))
+            throw new UserNotFoundException("User not found with id " + id);
+        
         userRepository.deletePermanently(id);
+
+        gatewayService.deleteUserPermanently(id);
     }
 
+    private void checkUniqueness(UserRequest userRequest) {
+        if (userRepository.existsByUsername(userRequest.getUsername()) || userRepository.existsByEmail(userRequest.getEmail())) {
+            throw new IllegalArgumentException("Username or email already in use");
+        }
+    }
+    private void checkUniquenessForUpdate(AppUser appUser, UserRequest userRequest){
+        if (!userRequest.getUsername().equals(appUser.getUsername()) && userRepository.existsByUsername(userRequest.getUsername())) {
+            throw new DataIntegrityViolationException("Username already exists: " + userRequest.getUsername());
+        }
+
+        if (!userRequest.getEmail().equals(appUser.getEmail()) && userRepository.existsByEmail(userRequest.getEmail())) {
+            throw new DataIntegrityViolationException("Email already exists: " + userRequest.getEmail());
+        }
+    }
+    private void validateRoles(List<String> roles) {
+
+        for (String role : roles) {
+            if (!DEFAULT_ROLES.contains(role)) {
+                throw new InvalidRoleException("Invalid role: " + role);
+            }
+        }
+    }
+    private AuthUserRequest mapToAuthUserRequest (UserRequest userRequest) {
+        return new AuthUserRequest(
+                userRequest.getUsername(),
+                userRequest.getPassword(),
+                userRequest.getRoles());
+    }
     private UserResponse mapToUserResponse(AppUser user) {
         return UserResponse.builder()
                 .id(user.getId())
@@ -124,7 +185,6 @@ public class UserServiceImpl implements UserService {
                 .roles(getRolesAsString(user.getRoles()))
                 .build();
     }
-
     private UserResponse mapToUserResponseWithoutPassword(AppUser user) {
         return UserResponse.builder()
                 .id(user.getId())
@@ -134,13 +194,11 @@ public class UserServiceImpl implements UserService {
                 .roles(getRolesAsString(user.getRoles()))
                 .build();
     }
-
     private List<String> getRolesAsString(List<Role> roles) {
         return roles.stream()
                 .map(Role::getRoleName)
                 .collect(Collectors.toList());
     }
-
     private List<Role> getRolesAsRole(List<String> roles) {
         List<Role> rolesList = new ArrayList<>();
         roles.forEach(roleName -> {
